refactor(agent-process): split agent_process.rs into submodules
Carve the 403-LOC single file into four cohesive submodules behind a
re-exporting `mod.rs`. No behavior change: every external import path
(`crate::agent_process::AgentProcesses`, `AgentWorker`,
`GLOBAL_AGENT_KEY`, `ensure_global_agent`, `retire_agent_key`,
`route_payload_key`, `write_agent_payload`, `project_root`) still
resolves, so `lib.rs`, `agent_commands.rs`, and the extension reload /
installer / watcher modules stay untouched.

Layout:

- `agent_process/sidecar.rs` — agent-binary path resolution
  (`project_root` for dev, `find_sidecar_binary` for release).
- `agent_process/spawn.rs` — `ensure_agent_spawned` plus the helper
  `build_command` / `apply_user_env` / `apply_worker_env` it now
  delegates to. Builds the dev `bun run agent/main.ts` vs release
  sidecar command, populates `AETHON_*` env, pipes stdio, hands the
  pipes to `readers.rs`.
- `agent_process/readers.rs` — the stdout / stderr reader threads.
  Stdout owns the producer side of the `mutation_routes` map and
  emits `agent-response` / `agent-reloaded` / `agent-crashed`; stderr
  emits `agent-stderr` and feeds the 32-line tail ring that the
  crash payload carries. Reader bodies are preserved verbatim aside
  from extracting their inputs into `StdoutReaderCtx` /
  `StderrReaderCtx` structs so the spawn callsite is no longer a
  six-argument positional contract.
- `agent_process/process.rs` — public surface: `AgentProcesses`,
  `AgentWorker`, `GLOBAL_AGENT_KEY`, `tab_agent_key`, plus
  `ensure_global_agent` / `retire_agent_key` / `write_agent_payload`
  / `route_payload_key`. `route_payload_key` is the consumer side of
  the `mutation_routes` map populated by `readers.rs`. The two
  colocated tests (`global_agent_key_is_stable`,
  `tab_agent_key_keeps_tab_prefix`) move with the code they
  exercise.

Verified: `cargo fmt`, `cargo check`, `cargo clippy -- -D warnings`,
`cargo test --lib` (158/158), `bunx tsc --noEmit`, `bunx eslint src/`
(0 errors / 5 pre-existing warnings), `bunx vitest run` (1124/1124),
`treefmt`.
