+++ src-tauri/src/agent_process/sidecar.rs
@@ -1,0 +1,64 @@
+//! Path resolution for the agent runtime.
+//!
+//! Two flavours: in dev we run `bun run agent/main.ts` from the project
+//! root (the directory holding `agent/main.ts`); in release we exec the
+//! bundled `aethon-agent` sidecar Tauri places next to the main binary.
+
+use std::path::{Path, PathBuf};
+
+/// Find the project root (the directory containing `agent/main.ts`). Tauri
+/// launches the dev binary with cwd set to `src-tauri/`, but our agent script
+/// lives one level up, so a naive relative path resolves to the wrong place.
+/// Walk up from cwd until we find the marker; fall back to cwd if nothing
+/// matches.
+pub(crate) fn project_root() -> PathBuf {
+    let cwd = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
+    let mut dir: &Path = &cwd;
+    for _ in 0..6 {
+        if dir.join("agent").join("main.ts").exists() {
+            return dir.to_path_buf();
+        }
+        match dir.parent() {
+            Some(p) => dir = p,
+            None => break,
+        }
+    }
+    cwd
+}
+
+/// Locate the bundled `aethon-agent` sidecar binary. Tauri's externalBin
+/// mechanism places sidecars next to the main executable on each platform
+/// (e.g. `Aethon.app/Contents/MacOS/aethon-agent-aarch64-apple-darwin`).
+/// Returns Err with a descriptive message when none of the candidate paths
+/// exist; the caller falls back to `bun run` in dev or surfaces the error
+/// in release.
+pub(super) fn find_sidecar_binary() -> Result<PathBuf, String> {
+    let exe = std::env::current_exe().map_err(|e| format!("current_exe: {e}"))?;
+    let exe_dir = exe
+        .parent()
+        .ok_or("current_exe has no parent dir")?
+        .to_path_buf();
+    let triple = env!("AETHON_TARGET_TRIPLE");
+    // Tauri's externalBin strips the triple suffix before placing the file
+    // alongside the main exe. Check the stripped variant first, then the raw
+    // triple form for direct target/release runs.
+    let ext = std::env::consts::EXE_SUFFIX;
+    let candidates = [
+        exe_dir.join(format!("aethon-agent{ext}")),
+        exe_dir.join(format!("aethon-agent-{triple}{ext}")),
+    ];
+    for path in &candidates {
+        if path.exists() {
+            return Ok(path.clone());
+        }
+    }
+    Err(format!(
+        "aethon-agent sidecar not found next to {} (looked for: {})",
+        exe.display(),
+        candidates
+            .iter()
+            .map(|p| p.display().to_string())
+            .collect::<Vec<_>>()
+            .join(", "),
+    ))
+}
